📝 docs(diff): 重构差异处理模块文档和实现
- 【重构】重写模块说明文档，提供更清晰的模块结构和使用说明
- 【优化】改进文档结构，增加详细的接口说明和使用示例
- 【删除】移除 DiffFormatter 旧文件
- 【重构】将 getDiffConfig 方法移至 types.ts
- 【新增】DiffSimplifier 增加空格压缩功能
- 【文档】补充了 DiffSplitter 和 DiffSimplifier 的完整使用示例
